feat: move markdown component in component library
